Update to 25.11 releases
# Changes

## Replace SDDM Sugar Candy theme with SDDM Astronaut

The documentation reports that the path to the background image must be
relative but I can't see any reason why in the source so we're going
to try absolute paths.

## Replace rofi-wayland with rofi

Support for Wayland has been merged into rofi upstream. Nixpkgs enables
wayland and x11 support by default but we only want Wayland so we disable
x11.

## Update config syntax

Options for some modules have changed.
- Git
- SSH
- Delta
- Nix GC
- Hyprland (upstream config has deprecated some values)